+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### Fixes
 
+- [#8714](https://github.com/blockscout/blockscout/pull/8714) - Fix sourcify check 
 - [#8705](https://github.com/blockscout/blockscout/pull/8705) - Fix sourcify enabled flag
 - [#8695](https://github.com/blockscout/blockscout/pull/8695) - Don't override internal transaction error if it's present already
 - [#8685](https://github.com/blockscout/blockscout/pull/8685) - Fix db pool size exceeds Postgres max connections